backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
